File statusCodes moved from Config folder into utils folder
