+++ Backend/src/controllers/user.controller.ts
@@ -4,7 +4,7 @@ import jwt from 'jsonwebtoken'; // Importing third party package
 import dotenv from 'dotenv'; // Importing third party package
 import { JWT_EXPIRES_IN, validateEmail } from '../utils/constants';
 import { MESSAGE_CODES } from '../utils/messageCodes'; // Importing responses from another file
-import { STATUS_CODES } from '../config/statusCodes'; // Importing status codes from another file
+import { STATUS_CODES } from '../utils/statusCodes'; // Importing status codes from another file
 
 dotenv.config();
